Trait festvox::TtsGenerator

pub trait TtsGenerator<'r>: Send + Sync + Sized + 'static {
    type Error: Debug;
    fn new() -> Result<Self, <Self as TtsGenerator<'r>>::Error>;

    fn generate<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        details: &'life1 PhrasePackage, 
        config: &'life2 Config
    ) -> Pin<Box<dyn Future<Output = Result<PathBuf, <Self as TtsGenerator<'r>>::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn fairing() -> AdHoc
    where
        &'r Self: FromRequest<'r>,
    { ... }
}

A trait indicating a tts generator that can be constructed and used to generate audio files from speech

Associated Types

type Error: Debug

Required methods

fn new() -> Result<Self, <Self as TtsGenerator<'r>>::Error>

Create a new TTS builder

fn generate<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
details: &'life1 PhrasePackage,
config: &'life2 Config
) -> Pin<Box<dyn Future<Output = Result<PathBuf, <Self as TtsGenerator<'r>>::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Generate a phrase utilising the TTS system, with the set parameters

Provided methods

fn fairing() -> AdHoc where
&'r Self: FromRequest<'r>,

Generate an adhoc fairing which can be bound to a launching rocket.